+++ CHANGES.md
@@ -2,6 +2,11 @@
 
 (next version)
 
+### Version 0.1.11 (2017-08-30)
+
+* Add metadata for sphinx
+* Add `convert(src)` function, which is shortcut of `m2r.M2R()(src)`
+
 ### Version 0.1.10 (2017-08-15)
 
 * Include CHANGES and test files in source distribution